+++ app/app.component.js
@@ -1,4 +1,4 @@
-import PartyMember from '../party-member';
+import PartyMember from '../lib/party-member';
 
 const AppComponent = new PartyMember('app-component', {}, {}, 'app/app.component.html');
