+++ app/date/date.component.js
@@ -1,4 +1,4 @@
-import PartyMember from '../../party-member';
+import PartyMember from '../../lib/party-member';
 
 const DateComponent = new PartyMember('date-component', { date: '2021-02-01' }, {}, 'app/date/date.component.html');
