+++ lib/build-system/bundler.js
@@ -2,9 +2,9 @@ import { build } from 'esbuild';
 
 export default function bundleJsModulesWithWatcher(minify) {
     build({
-        entryPoints: ['reactive/watcher.js'],
-        bundle: true,
+        entryPoints: ['lib/reactive/watcher.js'],
         outfile: 'public/bundle.js',
+        bundle: true,
         minify: minify,
         platform: 'node'
     })
