+++ lib/build-system/registrar.js
