+++ lib/build-system/templates.js
