+++ lib/build.js
@@ -1,19 +1,19 @@
 import fs from 'fs';
-import module from './index.module';
+import module from '../index.module';
 import Registrar from './build-system/registrar';
 import generateHtmlFromTemplate from './build-system/templates';
-import bundleJsModulesWithWatcher from './build-system/mini-bundle';
+import bundleJsModulesWithWatcher from './build-system/bundler';
+
+import { modulePath, debugFlag } from './cli/args';
 
 function build(module, template, debug) {
     for (const partyMember of module) {
         Registrar.registerPartyMember(partyMember);
     }
-
     const html = generateHtmlFromTemplate(template, debug);
     fs.writeFile('public/index.html', html, (err) => {
         if (err) throw err;
     });
-
     bundleJsModulesWithWatcher(!debug);
 }
 
@@ -21,5 +21,6 @@ fs.readFile('index.html', (err, template) => {
     if (err) {
         throw err;
     }
-    build(module, template, true);
+    const debug = debugFlag() === 'true';
+    build(module, template, debug);
 });
