+++ lib/cli/default-args.json
@@ -1,0 +1,4 @@
+{
+    "debug": true,
+    "module": null
+}
