+++ lib/party-member.js
